Version incomplète du 19/06/2018

+++ SIM/src/org/cazabat/sim/Constant.java
@@ -12,6 +12,6 @@ public class Constant {
 	public static final boolean SUCCESS=true;
 	public static final boolean ERROR=false;
 	
-	public static final int UNDEFINED=999;
+	public static final float UNDEFINED=999;
 	public static enum SYSTEM {LAMBERT93,WSG84};
 }
